# docs/generated/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/providers/stats": {
      "get": {
        "operationId": "ProviderManagementController_stats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Management"
        ]
      }
    },
...
    "/api/v1/providers/lookup": {
      "get": {
        "operationId": "ProviderManagementController_lookup",
        "parameters": [
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "approvalStatus",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Management"
        ]
      }
    },
    "/api/v1/providers/{id}": {
      "get": {
        "operationId": "ProviderManagementController_details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Management"
        ]
      },
      "patch": {
        "operationId": "ProviderManagementController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Management"
        ]
      }
    },
...
    "/api/v1/provider/inventory": {
      "get": {
        "operationId": "ProviderInventoryController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ALL",
                "ACTIVE",
                "INACTIVE",
                "OUT_OF_STOCK",
                "PENDING",
                "REJECTED"
              ]
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "name",
                "price",
                "stockQuantity"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Inventory"
        ]
      },
      "post": {
        "operationId": "ProviderInventoryController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderInventoryItemDto"
              },
              "examples": {
                "withVariants": {
                  "value": {
                    "name": "Luxury Perfume",
                    "description": "Long-lasting premium fragrance.",
                    "shortDescription": "Premium fragrance gift.",
                    "categoryId": "gift_category_id",
                    "price": 99.99,
                    "currency": "PKR",
                    "stockQuantity": 50,
                    "sku": "PERFUME-001",
                    "imageUrls": [
                      "https://cdn.yourdomain.com/gift-images/perfume.png"
                    ],
                    "isAvailable": true,
                    "variants": [
                      {
                        "name": "30ml",
                        "price": 89.99,
                        "originalPrice": 119.99,
                        "stockQuantity": 10,
                        "sku": "PERFUME-30ML",
                        "isPopular": false,
                        "isDefault": false,
                        "sortOrder": 1,
                        "isActive": true
                      },
                      {
                        "name": "50ml",
                        "price": 129.99,
                        "originalPrice": 159.99,
                        "stockQuantity": 20,
                        "sku": "PERFUME-50ML",
                        "isPopular": true,
                        "isDefault": true,
                        "sortOrder": 2,
                        "isActive": true
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Inventory"
        ]
      }
    },
    "/api/v1/provider/inventory/stats": {
      "get": {
        "operationId": "ProviderInventoryController_stats",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Inventory"
        ]
      }
    },
    "/api/v1/provider/inventory/lookup": {
      "get": {
        "operationId": "ProviderInventoryController_lookup",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Inventory"
        ]
      }
    },
    "/api/v1/provider/inventory/{id}": {
      "get": {
        "operationId": "ProviderInventoryController_details",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Inventory"
        ]
      },
      "patch": {
        "operationId": "ProviderInventoryController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderInventoryItemDto"
              },
              "examples": {
                "upsertVariants": {
                  "value": {
                    "replaceVariants": false,
                    "variants": [
                      {
                        "id": "variant_id",
                        "name": "50ml",
                        "price": 129.99,
                        "originalPrice": 159.99,
                        "stockQuantity": 20,
                        "sku": "PERFUME-50ML",
                        "isPopular": true,
                        "isDefault": true,
                        "sortOrder": 2,
                        "isActive": true
                      },
                      {
                        "name": "150ml",
                        "price": 249.99,
                        "originalPrice": 299.99,
                        "stockQuantity": 5,
                        "sku": "PERFUME-150ML",
                        "isPopular": false,
                        "isDefault": false,
                        "sortOrder": 4,
                        "isActive": true
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Inventory"
        ]
      },
      "delete": {
        "operationId": "ProviderInventoryController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Inventory"
        ]
      }
    },
    "/api/v1/provider/inventory/{id}/availability": {
      "patch": {
        "operationId": "ProviderInventoryController_updateAvailability",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderAvailabilityDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Provider Inventory"
        ]
      }
    },
...
  },
  "info": {
    "title": "Gift App Backend API",
    "description": "Gift App authentication and user account APIs",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateProviderDto": {
        "type": "object",
        "properties": {
          "businessName": {
            "type": "string",
            "example": "Gifts & Blooms Co. Ltd"
          },
          "phone": {
            "type": "string",
            "example": "+15551234567"
          },
          "serviceArea": {
            "type": "string",
            "example": "New York, USA"
          },
          "headquarters": {
            "type": "string",
            "example": "New York, USA"
          },
          "avatarUrl": {
            "type": "string",
            "example": "https://<YOUR_PUBLIC_BUCKET_OR_CDN_URL>/provider-logos/logo.png"
          },
          "documentUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ProviderInventoryVariantDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "variant_id"
          },
          "name": {
            "type": "string",
            "example": "50ml"
          },
          "price": {
            "type": "number",
            "example": 129.99
          },
          "originalPrice": {
            "type": "number",
            "example": 159.99
          },
          "stockQuantity": {
            "type": "number",
            "example": 20
          },
          "sku": {
            "type": "string",
            "example": "PERFUME-50ML"
          },
          "isPopular": {
            "type": "boolean",
            "example": true
          },
          "isDefault": {
            "type": "boolean",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "example": 2
          },
          "isActive": {
            "type": "boolean",
            "example": true
          }
        },
        "required": [
          "name",
          "price",
          "stockQuantity"
        ]
      },
      "CreateProviderInventoryItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "stockQuantity": {
            "type": "number"
          },
          "sku": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isAvailable": {
            "type": "boolean"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderInventoryVariantDto"
            }
          }
        },
        "required": [
          "name",
          "price",
          "categoryId"
        ]
      },
      "UpdateProviderInventoryItemDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "shortDescription": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "currency": {
            "type": "string"
          },
          "stockQuantity": {
            "type": "number"
          },
          "sku": {
            "type": "string"
          },
          "categoryId": {
            "type": "string"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isAvailable": {
            "type": "boolean"
          },
          "replaceVariants": {
            "type": "boolean",
            "example": false
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProviderInventoryVariantDto"
            }
          }
        }
      },
      "UpdateProviderAvailabilityDto": {
        "type": "object",
        "properties": {
          "isAvailable": {
            "type": "boolean"
          }
        },
        "required": [
          "isAvailable"
        ]
      },
...
    }
  }
}